Initial Commit
This commit contains quite a bit of work. It has a basic implementation of TAMusicScore, TAMusicPart, and TAMusicMeasure. It also contains unit tests and the first beginnings of the TAMusicXMLImporter.

This also contains the MusicXML Test Cases. These are mainly for visual usage, but provide a great base for different xml files.